split example package
