+++ src/test/java/com/esaulpaugh/headlong/abi/example/ABIStudent.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.esaulpaugh.headlong.example;
+package com.esaulpaugh.headlong.abi.example;
 
 import com.esaulpaugh.headlong.abi.Tuple;
